# api/docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  domain.Course:
    properties:
      code:
        type: string
      color:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      name:
        type: string
      published:
        type: boolean
      updatedAt:
        type: string
    type: object
...
  domain.Offer:
    properties:
      benefits:
        items:
          type: string
        type: array
      description:
        type: string
      id:
        type: string
      name:
        type: string
      packages:
        items:
          type: string
        type: array
      paymentMethod:
        $ref: '#/definitions/domain.PaymentMethod'
      price:
        $ref: '#/definitions/domain.Price'
      schoolId:
        type: string
    type: object
...
  domain.PaymentMethod:
    properties:
      provider:
        type: string
      usesProvider:
        type: boolean
    type: object
  domain.Price:
    properties:
      currency:
        type: string
      value:
        type: integer
    type: object
  domain.PromoCode:
    properties:
      code:
        type: string
      discountPercentage:
        type: integer
      expiresAt:
        type: string
      id:
        type: string
      offerIds:
        items:
          type: string
        type: array
      schoolId:
        type: string
    type: object
...
  v1.createCourseInput:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  v1.dataResponse:
    properties:
      count:
        type: integer
      data:
        type: object
    type: object
...
  v1.idResponse:
    properties:
      id:
        type: object
    type: object
...
  v1.response:
    properties:
      message:
        type: string
    type: object
...
  v1.updateCourseInput:
    properties:
      color:
        type: string
      description:
        type: string
      imageUrl:
        type: string
      name:
        type: string
      published:
        type: boolean
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: REST API for UltraThreads App
  title: UltraThreads API
  version: "1.0"
paths:
...
  /admins/courses:
    get:
      consumes:
      - application/json
      description: admin get all courses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.dataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Get All Courses
      tags:
      - admins-courses
    post:
      consumes:
      - application/json
      description: admin create new course
      parameters:
      - description: course info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.createCourseInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.idResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Create New Courses
      tags:
      - admins-courses
  /admins/courses/{id}:
    delete:
      consumes:
      - application/json
      description: admin delete course
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Delete Course
      tags:
      - admins-courses
    get:
      consumes:
      - application/json
      description: admin get course by id
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Get Course By ID
      tags:
      - admins-courses
    put:
      consumes:
      - application/json
      description: admin update course
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      - description: course update info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/v1.updateCourseInput'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - AdminAuth: []
      summary: Admin Update Course
      tags:
      - admins-courses
...
  /courses:
    get:
      consumes:
      - application/json
      description: get all courses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.dataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get All Courses
      tags:
      - courses
  /courses/{id}:
    get:
      consumes:
      - application/json
      description: get course by id
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Course'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get Course By ModuleID
      tags:
      - courses
  /courses/{id}/offers:
    get:
      consumes:
      - application/json
      description: get course offers
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.dataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get Course Offers
      tags:
      - courses
  /offers/{id}:
    get:
      consumes:
      - application/json
      description: get offer by id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Offer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get Offer By ID
      tags:
      - offers
  /promocodes/{code}:
    get:
      consumes:
      - application/json
      description: get promocode by code
      parameters:
      - description: code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.PromoCode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
        default:
          description: ""
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get PromoCode By Code
      tags:
      - promocodes
...
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  StudentsAuth:
    in: header
    name: Authorization
    type: apiKey
  UsersAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"